backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
celery==5.3.6
redis==5.0.1
yt-dlp==2023.11.16
reportlab==4.0.9
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
chromadb==0.4.22
gunicorn==21.2.0
python-docx==1.1.0
python-pptx==0.6.23
faster-whisper==0.10.0
groq==0.4.2
python-dotenv==1.0.0
